tests/wpt/tests/html/canvas/element/drawing-images-to-the-canvas/2d.drawImage.negativedir.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.drawImage.negativedir</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.drawImage.negativedir</h1>
<p class="desc">Negative dimensions do not affect the direction of the image</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
promise_test(async t => {

  var canvas = document.getElementById('c');
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  const response = await fetch('/images/ggrr-256x256.png');
  const blob = await response.blob();
  const bitmap = await createImageBitmap(blob);
  ctx.drawImage(bitmap, 0, 178, 50, -100, 0, 0, 50, 100);
  ctx.drawImage(bitmap, 0, 78, 50, 100, 50, 100, 50, -100);
  _assertPixelApprox(canvas, 1,1, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 1,48, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 98,1, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 98,48, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 48,1, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 48,48, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 51,1, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 51,48, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 25,25, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 75,25, 0,255,0,255, 2);

}, "Negative dimensions do not affect the direction of the image");
</script>
